src/app/Services/BookService/book.service.ts
import { Injectable } from '@angular/core';
import { HttpService } from '../HttpService/http.service';
import { HttpHeaders } from '@angular/common/http';
import { Observable } from 'rxjs';

@Injectable({
  providedIn: 'root'
})
export class BookService {

  token: any;
  constructor(private httpService: HttpService) { }

  getAllBooksData(): Observable<any> {
    this.token = localStorage.getItem('token');

    let httpHeadersOption = {
      headers: new HttpHeaders({
        contentType: 'application/json',
        authorization: this.token
      })
    }

    return this.httpService.getService("get/book", false, httpHeadersOption)
  }
}
